requirements.txt
# Core dependencies
pdfplumber>=0.10.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
groq>=0.4.0
streamlit>=1.28.0
torch>=2.0.0  # PyTorch backend for sentence-transformers (avoids TensorFlow issues)

# Optional dependencies (for fallback PDF extraction)
pdfminer.six>=20221105

# ML dependencies
xgboost>=2.0.0
scikit-learn>=1.3.0
shap>=0.42.0
pandas>=2.0.0
joblib>=1.3.0

# Utilities
numpy>=1.24.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0

